Staging

+++ app/Http/Controllers/TaskController.php
@@ -16,6 +16,8 @@ use App\Models\Fortnight;
 use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\View\View;
+use Illuminate\Support\Facades\Mail;
+use App\Mail\TaskAssigned;
 
 class TaskController extends Controller
 {
@@ -27,44 +29,76 @@ class TaskController extends Controller
             $headOf = request()->user()->load('headOf')->headOf;
 
             $tasks = $headOf ? $headOf->tasks() : Task::query();
-            
         } else if (request()->user()->hasAnyRole(['SUPER_ADMIN', 'ADMIN'])) {
 
             $tasks = Task::with(['target', 'departments']);
         } else {
-
             $tasks = request()->user()->tasks();
         }
 
-        /** 
-         * Fetch currentFortnight by checking to which fortnight the current day belongs to(we need it to fetch current)
-         * 
-         * we need the currentFortnight variable to fetch tasks of currentFortnight and 
-         * 
-         * to create Day just in case the current day doesnt exist
-         * 
-         * */
+        // Fetch tasks only related to the currently active fortnight if currentFortnight is true
+        $currentFortnight = null;
+
         $today = Carbon::now()->format('Y-m-d');
+        if ($request->query('currentFortnight')) {
 
-        $currentFortnight = $request->query('currentFortnight') || $request->query('onlyToday') ? Fortnight::whereDate('start_date', '<=', $today)
-            ->whereDate('end_date', '>=', $today)->first() : null;
 
-        // Only if The URL contains the currentFortnight or onlyToday
-        if ($request->query('currentFortnight') || $request->query('onlyToday')) {
+            $currentFortnight = Fortnight::whereDate('start_date', '<=', $today)
+                ->whereDate('end_date', '>=', $today)->first();
 
-            /**
-             * 
-             * Filter by tasks by scopes 
-             * Scopes refer to how long the task takes to complete...(for now)
-             * 
-             */
-            $tasks = $this->filterByScope($tasks, $request, $currentFortnight, $today);
+            $tasks = $tasks->whereHas('fortnights', function ($query) use ($currentFortnight) {
+                $query->where('fortnights.id', $currentFortnight->id);
+            });
+        } elseif ($request->query('todays')) {
+            $currentFortnight = Fortnight::whereDate('start_date', '<=', $today)
+                ->whereDate('end_date', '>=', $today)->first();
+
+            if (Day::where('date', $today)->exists()) {
+                $date = Day::where('date', $today)->first()->id;
+            } else {
+                $date = Day::create([
+                    'fortnight_id' => $currentFortnight->id,
+                    'date' => $today
+                ])->id;
+            }
+
+            $tasks = $tasks->whereHas('days', function ($query) use ($date) {
+                $query->where('days.id', $date);
+            });
+
+            $currentFortnight = null;
         }
 
-        // This Filters the tasks as the user requested
-        $tasks = $this->filterByColumns($tasks, $request);
+        // Filtering and sorting
+        $search = $request->query('search') ?? null;
+        $status = $request->query('status') ?? null;
+        $dueDays = is_numeric($request->query('due_days')) ? intval($request->query('due_days')) : null;
+        $order = strtolower($request->query('order'));
 
-        $tasks = $tasks->with(['target', 'departments', 'createdBy'])->where('is_subtask', false)->paginate(15);
+        if ($search) {
+            $tasks = $tasks->where(function ($query) use ($search) {
+                $query->where('name', 'LIKE', '%' . $search . '%')
+                    ->orWhereHas('createdBy', function ($q) use ($search) {
+                        $q->where('name', 'LIKE', '%' . $search . '%');
+                    });
+            });
+        }
+
+        if ($status) {
+            $tasks = $tasks->where('status', $status);
+        }
+
+        if ($dueDays) {
+            $dueDate = Carbon::now()->addDays($dueDays);
+            $tasks = $tasks->whereNotNull('due_date')->whereDate('due_date', '<=', $dueDate);
+        }
+
+        if (in_array($order, ['asc', 'desc'])) {
+            $tasks = $tasks->orderBy('name', $order);
+        }
+
+
+        $tasks = $tasks->with(['target', 'departments', 'createdBy'])->paginate(10);
 
         return view('tasks.index', compact('tasks', 'currentFortnight'));
     }
@@ -77,9 +111,9 @@ class TaskController extends Controller
 
         $parent_tasks = Task::get();
 
-        $users = request()->user()->hasROle('DEPARTMENT_HEAD') ? request()->user()->headOf->users : (request()->user()->hasRole('EMPLOYEE') ? [] : User::get());
+        $users = request()->user()->hasROle('DEPARTMENT_HEAD') ? request()->user()->headOf->users :  (request()->user()->hasRole('EMPLOYEE') ? [] : User::get());
 
-        $forToday = (bool) request()->query('dailyTask') === true ? Carbon::now()->format('Y-m-d') : null;
+        $today = (bool) request()->query('dailyTask') === true ? Carbon::now()->format('Y-m-d') : null;
 
         $fortnights = Fortnight::get();
 
@@ -87,75 +121,118 @@ class TaskController extends Controller
 
         $parent_task_id = $request->input('parent_task_id', null);
 
-        return view('tasks.create', compact('targets', 'parent_tasks', 'departments', 'users', 'fortnights', 'assignedUsers', 'parent_task_id', 'forToday'));
+        $forToday = null;
+
+        return view('tasks.create', compact('targets', 'parent_tasks', 'departments', 'users', 'fortnights', 'assignedUsers', 'parent_task_id', 'today', 'forToday'));
     }
 
     public function store(TaskStoreRequest $request)
     {
         $data = $request->validated();
 
+        if (isset($data['parent_task_id'])) {
+
+            $data['is_subtask'] = true;
+
+        } else {
+
+            $data['is_subtask'] = false;
+        }
+
+    
         $data['is_subtask'] = $data['parent_task_id'] ? true : false;
+
         $data['created_by'] = request()->user()->id;
 
         /**
          * Sets assigned departments ($departments) to the department the logged in user belongs to if the role of the user is EMPLOYEE or DEPARMENT_HEAD 
          * But sets the assigned departments to what the logged in user requested to assign if the role is any other than the above(^)
          * 
-         * sets the assigned users ($users) to the logged in user if the role is employee or sets it to what the logged in user assigned if the role is otherwise
+         * sets the assigned users ($users) to the logged in user if the role is employee or sets it to the users the logged in user assigned if the role is otherwise
          */
+    
         $departments = request()->user()->hasAnyRole(['EMPLOYEE', 'DEPARTMENT_HEAD']) ? [0 => request()->user()->department->id] : $request['department_id'];
+
         $fortnights = $request['fortnight_id'];
+
         $users = request()->user()->hasRole('EMPLOYEE') ? [0 => request()->user()->id] : $request['user_id'];
 
         // unsets(removes) fields from the $data array which dont belong to the tasks table
+    
         unset($data['department_id']);
+
         unset($data['fortnight_id']);
+
         unset($data['user_id']);
 
+    
         $task = Task::create($data);
-
         /** 
          * If the task created is daily the client form will send a query called forToday in the url
          * 
          * in that case we will fetch the current day from the database and attach it to the task that was just created and if the day doesnt exist we will create it
          * 
          */
+    
         if ($request->query('forToday')) {
 
             $today = Carbon::now()->format('Y-m-d');
 
+    
             $currentFortnightId = Fortnight::whereDate('start_date', '<=', $today)
                 ->whereDate('end_date', '>=', $today)->first()->id;
 
+    
             if (Day::where('date', $today)->exists()) {
                 $day = Day::where('date', $today)->first();
+
             } else {
                 $day = Day::create([
                     'fortnight_id' => $currentFortnightId,
                     'date' => $today
                 ]);
             }
-
+    
             $day->tasks()->attach($task);
         }
-
+        if (!request()->user()->hasRole('EMPLOYEE')) {
+            $task->departments()->attach($departments);
+        }
+        else {
+            $task->departments()->attach($departments);
+        }
+        
         /**
          * Attaches the departments, users and fortnights that were created above to the newly created task
          */
-
+    
         $task->departments()->attach($departments);
+
         $task->fortnights()->attach($fortnights);
+
         $task->users()->attach($users);
 
+    
+        // Get the assigning user
+        $assigningUser = request()->user();
+    
+        // Send email to assigned users
+        foreach ($users as $userId) {
+
+            $user = User::find($userId);
+
+            Mail::to($user->email)->send(new TaskAssigned($task, $assigningUser));
+        }
+    
         return redirect()->route('tasks.index')->with('status', 'Task has been successfully created.');
     }
-
+    
     public function show(Task $task)
     {
         $task = $task->load('departments', 'users', 'kpis', 'subtasks');
 
-        $users = $task->users()->paginate(15);
-        
+        $users = $task->users;
+
         return view('tasks.show', compact('task', 'users'));
     }
 
@@ -169,7 +246,7 @@ class TaskController extends Controller
 
         $isDaily = $task->days()->exists();
 
-        $forToday = $isDaily ? $task->days()->first() : null;
+        $today = $isDaily ? $task->days()->first() : null;
 
         $assignedUsers = $task->users()->pluck('users.id')->toArray();
 
@@ -179,36 +256,34 @@ class TaskController extends Controller
 
         $parent_tasks = Task::where('is_subtask', false)->get();
 
-        $users = request()->user()->hasROle('DEPARTMENT_HEAD') ? request()->user()->headOf->users : (request()->user()->hasRole('EMPLOYEE') ? [] : User::get());
+        $users = User::get();
 
         $fortnights = Fortnight::get();
 
-        return view('tasks.edit', compact('task', 'targets', 'parent_tasks', 'departments', 'users', 'fortnights', 'assignedUsers', 'forToday'));
+        $forToday = null;
+
+        return view('tasks.edit', compact('task', 'targets', 'parent_tasks', 'departments', 'users', 'fortnights', 'assignedUsers', 'today', 'forToday'));
     }
 
     public function update(TaskUpdateRequest $request, Task $task)
     {
-        /**
-         * The manageTask is a custom made permission to check if the employee, department_head can affect the task (TaskPolicy.php) 
-         */
         if (request()->user()->cannot('manageTask', $task)) {
             abort(403);
         }
-        /**
-         * Most of the logic is listed in the store method
-         */
 
         $data = $request->validated();
+
         $data['is_subtask'] = $data['parent_task_id'] ? true : false;
 
-        $departments = request()->user()->hasAnyRole(['EMPLOYEE', 'DEPARTMENT_HEAD']) ? [0 => request()->user()->department->id] : $request['department_id'];
+        $departments = $request['department_id'];
         $fortnights = $request['fortnight_id'];
         $users = request()->user()->hasRole('EMPLOYEE') ? [0 => request()->user()->id] : $request['user_id'];
 
-        //Detach the existing relations with this 3 models and this this task in order to reset them based on what the user has adjusted
-        $task->departments()->detach();
-        $task->fortnights()->detach();
-        $task->users()->detach();
+        if (!request()->user()->hasRole('EMPLOYEE')) {
+            $task->departments()->detach();
+            $task->fortnights()->detach();
+            $task->users()->detach();
+        }
 
 
         unset($data['department_id']);
@@ -217,9 +292,11 @@ class TaskController extends Controller
 
         $task->update($data);
 
-        $task->departments()->attach($departments);
-        $task->fortnights()->attach($fortnights);
-        $task->users()->attach($users);
+        if (!request()->user()->hasRole('EMPLOYEE')) {
+            $task->departments()->attach($departments);
+            $task->fortnights()->attach($fortnights);
+            $task->users()->attach($users);
+        }
 
         return redirect()->route('tasks.index')->with('status', 'Task has been successfully Updated.');
     }
@@ -232,7 +309,7 @@ class TaskController extends Controller
 
         if ($task->kpis()->exists() || $task->deliverables()->exists() || $task->feedbacks()->exists()) {
             return redirect()->route('tasks.index')
-                ->with('related', 'You can\'t Delete This pending Task it have feedback on it.');
+                ->with('status', 'You can\'t Delete This pending Task it have feedback on it.');
         }
 
         $task->delete();
@@ -241,32 +318,35 @@ class TaskController extends Controller
     }
 
     public function listByStatus($status)
-    {
+    {        
         $tasks = Task::where('status', $status)->get();
 
-        if (count($tasks) > 0) {
+        if(count($tasks) > 0)
+        {
 
             if (request()->user()->hasAnyRole(['DEPARTMENT_HEAD'])) {
 
                 $headOf = request()->user()->load('headOf')->headOf;
 
                 if ($headOf) {
-                    $tasks = $headOf->tasks()->with(['target', 'departments'])->where('status', $status)->paginate(15);
+                    $tasks = $headOf->tasks()->with(['target', 'departments'])->where('status', $status)->paginate(10);
                 } else {
-                    $tasks = Task::with(['target', 'departments'])->where('status', $status)->paginate(15);
+                    $tasks = Task::with(['target', 'departments'])->where('status', $status)->paginate(10);
                 }
+            
             } else if (request()->user()->hasAnyRole(['SUPER_ADMIN', 'ADMIN'])) {
-
-                $tasks = Task::with(['target', 'departments'])->where('status', $status)->paginate(15);
+                
+                $tasks = Task::with(['target', 'departments'])->where('status', $status)->paginate(10);
+            
             } else {
-
-                $tasks = request()->user()->tasks()->with(['target', 'departments'])->where('status', $status)->paginate(15);
+                
+                $tasks = request()->user()->tasks()->with(['target', 'departments'])->where('status', $status)->paginate(10);
             }
 
             $currentFortnight = null;
 
             return view('tasks.index', compact('tasks', 'currentFortnight'));
-        } else {
+        }else{
             return redirect()->route('index');
         }
     }
@@ -282,97 +362,5 @@ class TaskController extends Controller
         $task->save();
 
         return redirect()->route('tasks.show', $task->id)->with('success', 'Task status updated successfully.');
-    }
-
-
-    /**
-     * 
-     * Filters tasks base on the received $tasks and $requested
-     * 
-     * this function can filter using search, status and due_date
-     * 
-     */
-    public function filterByColumns($tasks, $request)
-    {
-
-        $search = $request->query('search') ?? null;
-
-        $status = $request->query('status') ?? null;
-
-        $dueDays = is_numeric($request->query('due_days')) ? intval($request->query('due_days')) : null;
-
-        $order = strtolower($request->query('order'));
-
-        //Search Filter
-        if ($search) {
-
-            // This is a grouped query that lets the function search from task name and also their creator name
-            $tasks = $tasks->where(function ($query) use ($search) {
-
-                $query->where('name', 'LIKE', '%' . $search . '%')
-                    ->orWhereHas('createdBy', function ($q) use ($search) {
-                        $q->where('name', 'LIKE', '%' . $search . '%');
-                    });
-            });
-        }
-
-        // This filters The tasks based on Status
-        if ($status) {
-            $tasks = $tasks->where('status', $status);
-        }
-
-        // This filters based on how many days is left before task expires (based on due date)
-        if ($dueDays) {
-            $dueDate = Carbon::now()->addDays($dueDays)->toDateString();
-            $tasks = $tasks->whereNotNull('due_date')->whereDate('due_date', '<=', $dueDate);
-        }
-
-        // This adjusts the orders based on the name of the tasks
-        if (in_array($order, ['asc', 'desc'])) {
-            $tasks = $tasks->orderBy('name', $order);
-        }
-
-        // Return The query builder to the caller
-        return $tasks;
-    }
-
-    /**
-     * 
-     * This Function is what enables us to filter currentFortnights or Daily Tasks
-     * 
-     */
-
-    public function filterByScope($tasks, $request, $currentFortnight, $today)
-    {
-        /**
-         * If There is ?currentFortnight=1 in the url we fetch the tasks of the current fortnight
-         * If There is ?onlyToday=1 in the url we fetch the tasks of the current day
-         */
-        if ($request->query('currentFortnight')) {
-
-            $currentFortnight = Fortnight::whereDate('start_date', '<=', $today)
-                ->whereDate('end_date', '>=', $today)->first();
-
-            $tasks = $tasks->whereHas('fortnights', function ($query) use ($currentFortnight) {
-                $query->where('fortnights.id', $currentFortnight->id);
-            });
-        } elseif ($request->query('onlyToday')) {
-
-            // Check if the day exists and if it doesnt create the day
-            if (Day::where('date', $today)->exists()) {
-                $day = Day::where('date', $today)->first()->id;
-            } else {
-                $day = Day::create([
-                    'fortnight_id' => $currentFortnight->id,
-                    'date' => $today
-                ])->id;
-            }
-            $tasks = $tasks->whereHas('days', function ($query) use ($day) {
-                $query->where('days.id', $day);
-            });
-        }
-
-        // Return the query builder
-        return $tasks;
     }
 }
